hvc: make HVC_ALLOC_TTY_ADAPTERS configurable
HVC_ALLOC_TTY_ADAPTERS determines how many hvc dev nodes are created,
and was hardcoded to 64. This can introduce some unnecessary delay
in the bootup when systemd scans and creates units for all nodes.
In such cases this number should be reduced, so made a config entry
which defaults to the original value of 64 to avoid unexpected change.
Set a range from 1 to 64 to allow it to be reduced to the minimum valid
amount.

Change-Id: I605d8af857a45503e77a02e990b54f366e4c8606
Signed-off-by: Johnathan Cobden-Nolan <quic_jcobdenn@quicinc.com>