arm: defconfig: sa525m: reduce HVC_ALLOC_TTY_ADAPTERS
The hvc nodes are not needed for televm, and the default of 64
nodes results in slower boot KPIs, reducing the count to the
minimum of 1, which is required for console connected to hostvm.
Do the same for fotavm. Reduce number to 10 for hostvm, since we
currently use 5 and 6 for the GVMs and leave room for a few more.

Change-Id: Ibf970e596754c7133331f8ac73a8572590a0d558
Signed-off-by: Johnathan Cobden-Nolan <quic_jcobdenn@quicinc.com>