usb: dwc3: Update dwc3 suspend flag to allow USB functions suspend
The USB functions are entered into the suspend state when USB runtime
suspend is enabled on the host side. This is achieved by receiving the
dwc3 gadget suspend interrupt on the dwc3 USB driver. In this scenario,
the USB functions can be resumed at any time through the wakeup source
if the functions have the remote wakeup capability, and then the USB
functions can go back to the suspend state when they receive the gadget
suspend interrupt. There is a possibility that the functions will not
enter suspend again even if the gadget suspend interrupt is received.
This is due to the dwc3 suspended flag is not updated if the gadget
has previously resumed. Fix this issue by updating the dwc3 suspended
flag in the dwc3 gadget resume sequence.

Change-Id: I7b35209d54df540c1dd53bd52630bafc56cf063e
Signed-off-by: Selvarasu Ganesan <quic_selvaras@quicinc.com>