stmmac: Fix compilation issue for auto target
lts tag android13-5.15_144 introduced the dependency
for DWMAC_LOONGSON on MACH_LOONGSON64
hence this configuration brought up the difference between
gki and msm gen3auto configs
since STMMAC is being enabled for gen3auto.

For mobile targets these all net drivers are disabled
so reverted this dependency as it is not NA for auto targets also.

Change-Id: I0a1195cba331cc5eaa050195fc5869492ff9fed3
Signed-off-by: Sivasri Kumar, Vanka <quic_svanka@quicinc.com>