commit bbc7118b1704 ("drivers: iio: Reduce sleep time for asm330lhh reset")
Reduce sleep time to 40ms, after asm330lhh reset. The turn on time
for asm330lhh is 35ms based on the datasheet. Add logic to wait until
the register is reset, and replace sleep with this for memory content
reboot. This change is to reduce proe time, thereby improving boot KPI.

Change-Id: I5e6d16431bd79b7cafdd651b4a41d022d1ec6113
Signed-off-by: Saurav Shah <quic_sauravs@quicinc.com>