defconfig: sa525m: enable IPC logging with procfs and bootparam support
IPC logging now supports being enabled without debugfs, and has an added
procfs interface for retrieving the logs. Enable this on sa525m which
has debugfs disabled on perf builds.

IPC logging also has added feature to be activated only when a parameter
is added to the kernel command line. Enable this on sa525m to avoid the
memory cost in perf builds.

Change-Id: Ic7e42df5b1e7a982794b5c9b7cd01c9fb428d4c2
Signed-off-by: Johnathan Cobden-Nolan <quic_jcobdenn@quicinc.com>