serial: msm_geni_serial: Add a config based check to avoid error print
For the perf build, when the UART is being probed, error logs are
observed, indicating a failure to create a debug directory.
To address this issue, a condition check for CONFIG_IPC_LOGGING has
been added in the debug initialization.
This is because the CONFIG_IPC_LOGGING configuration is disabled for
the performance build.

Change-Id: I6720e3b3a08cc363fbae49bc70f71f973827a577
Signed-off-by: Zahir Shabbir Khan <quic_zshabbir@quicinc.com>
(cherry picked from commit 0063785457df3b175a25844c68492850aab52787)