msm: mhi_dev: Update state info if M0 polling fails
The change <aae5affe7cf7> ("msm: mhi_dev: Synchronize D3_COLD handling
and MHI reset") is added to support MHI reset before D3cold in Linux
host. Here in d3cold after MHI reset, mhi_dev_enable() is interrupted
and is queued as part of next D0 event. In current implementation,
mhi_dev_enable() is requeued as part of D0 handling if MHI state is
found to be in disconnected state.

In windows host case, MHI reset is not received prior to D3cold and
mhi_dev_enable() is not interrupted. But if MHI reset is received during
D0 when MHI host driver is going through disable to enable state, it is
resulting in queueing of mhi_dev_enable() since MHI state is found to be
disconnected state. But queueing of this work is not required in this
case. This is causing a deadlock between MHI reset received and
mhi_dev_enable() work that is queued.

To resolve this, updating the state info as invalid if M0 polling fails
in mhi_dev_enable() and changing the check to queue mhi_dev_enable()
from D0 based on this.

Change-Id: I83acc311cb6aa27079f9f86ef339296fa90d0f17
Signed-off-by: Sai Chaitanya Kaveti <quic_skaveti@quicinc.com>