msm: ep_pcie: fix link reset issue in BME unset case
With commit c418ea261ed6 ("msm: ep_pcie: Do BCR reset
before link re-training"), a regression was introduced
in cases where BME is not set and Link was up, we will
still go for BCR reset. In these cases we should directly
jump to check BME instead of retraining the link as it
was already established.

Issue Flow:- enable_endpoint is being called from probe
as well as perst-deassert. The link will be up from the
first call but host takes some time to set the BME. When
the second instance is scheduled it will fail the enabled
link check and then we would go for BCR and relink up
which is not an ideal case.

Change-Id: I7588132a92b2dfb380bb7d219e25d45c940490d4
Signed-off-by: Nitesh Gupta <quic_nitegupt@quicinc.com>