msm: ep_pcie: Fix checkbme label
The checkbme label is intended to direct the flow to the code
that checks if the BME is set by the host. Previously, it was
used in the FW (PBL/XBL) linkup scenario to jump to the
BW scaling code. This change corrects the labels to ensure they
direct to the appropriate path.

This change extends the commit 4ae342a6eb0e ("msm: ep_pcie: fix
link reset issue in BME unset case"), which addresses a regression
introduced by commit c418ea261ed6 ("msm: ep_pcie: Do BCR reset
before link re-training").

Change-Id: I0b3a99f28692466605ce0c4d4645296a5de29fe4
Signed-off-by: Nitesh Gupta <quic_nitegupt@quicinc.com>