usb: gadget: u_serial: Replace list_del with list_del_init
Replace list_del() in some functions to list_del_init() because
list_del is supposed to be used when the given entry is removed
from the list. If the entry might ever be reinserted into another
list, we should use list_del_init, which reinitializes the linked
list pointers. This would prevent list corruptions.

Change-Id: Ie2baea07c8f59af062b7dc714777daa9e176240f
Signed-off-by: Prashanth K <quic_prashk@quicinc.com>