msm: mhi_dev: Defer creation of /dev/mhi_ctrl till link is up
The pcie probe returns success when in usb mode to meet clock driver
sync state requirement, this change was added with commit
<5864b01e3d40> ("msm: ep_pcie: Sync state failure in usb mode")

Making the ep-pcie driver probe successful allows the MHI-dev driver to
probe, as it is dependent on ep-pcie. This change was added by commit:
<a3c333ba079f> (“ARM: dts: msm: Allow MHI driver probe only when ep-pcie
driver is successfully probed”) and creates the /dev/mhi-ctrl node,
which is used by some user-space clients to identify if the transport
medium is PCIe or USB.

To avoid creating /dev/mhi-ctrl even when the actual transport is not
PCIe, defer the UCI initialization part from probe until linkup.
Do the mhi_uci_init after linkup received from ep_pcie.

Change-Id: I7c1091a031765150f6713c8810c1471112ce9bd9
Signed-off-by: Sumit Kumar <quic_sumk@quicinc.com>