msm: ep_pcie: Restructure the vreg init API
The vreg init and deinit is done every time the device follow the
suspend resume path, These APIs set the voltage and load on
every call which results in increased latency in the process path
as it involves communicating with another processor to set actual
voltage and load.

Change the process paths set the voltage and load only once when
the device is probed.
When the device is enabled only call the vreg_enable API
instead of setting the voltage and load every time.

Change-Id: Ib61fe4d888c2009cd4f937d309a283298df5c08b
Signed-off-by: Sumit Kumar <quic_sumk@quicinc.com>