qmi_interface: Update all dynamic allocations to vzalloc
Optional fields in QMI payload may not be present and decoding will
not happen. In such case, destination structure elements which are
part of optional fields should be initialized to zero to avoid
garbage values in it.

Update all dynamic allocations to use vzalloc instead of vmalloc
to initialize memory to zero.

Change-Id: I555800402345d61077eee474db72c7854d0f5da9
Signed-off-by: Vishnu Santhosh <quic_vishsant@quicinc.com>