mhi: bus: Set the re-queue data buffer size based on buffer len
The size of data packet received on EP can exceed MTU due to the
additional header based on the interface. Example for Ethernet-type
packets, the size of data packet received on EP is {MTU + Ethernet
header (18 bytes)}. For more details, refer to the documentation:
https://docs.kernel.org/networking/netdevices.html

Based on the current logic, the size of the re-queue data buffers
that RC queues to the EP is determined by the smaller value between
MTU and MRU. In cases where the size is set to the MTU, the packet
size received on the EP (MTU + Ethernet header) is larger than the
data buffer size (MTU) queued by the RC onto the EP.

Although EP driver has logic to break a large packet into smaller
chunks, EP provides separate callbacks for each of these smaller
chunks to client for the same packet, leading to a crash in
mhi_dev_net layer.

Change-Id: I834853ce60d6cf980d5fd7a614a0f8468655d528
Signed-off-by: Naga Rashmi Ayiluri <quic_nayiluri@quicinc.com>