usb: dwc3: dwc3-msm-ops: Replace DWC3 CONNDONE interrupt kretprobe
Since the compiler has removed the dwc3_gadget_conndone_interrupt() as its
own label (and has inlined the API), kretprobes can't be utilized.
Instead, replace the conndone kretprobe with  __dwc3_gadget_ep_enable()
and explicitly check for DWC3_DEPCFG_ACTION_MODIFY.  The action is set to
the modify action, as the DEP parameters need to be updated after finding
out the link speed.  This action is used only after receiving the CONNDONE
interrupt.

Change-Id: Ie00044aa0ee0b9c5c954b668d3c6a19ec7757039
Signed-off-by: Wesley Cheng <quic_wcheng@quicinc.com>
Signed-off-by: Prashanth K <quic_prashk@quicinc.com>