power: reset: qcom-reboot-reason: store panic and watchdog bark reasons
Reset due to panic does not call reboot notifiers, so add a panic
notifier which explicitly stores the panic reason, allowing this
reason to be seen after reset.

Watchdog bark was being stored when CONFIG_POWER_RESET_QCOM_RESET_REASON
was enabled, which is not the case with the new
CONFIG_POWER_RESET_QCOM_REBOOT_REASON_BOOTPARAM feature. Create feature
CONFIG_POWER_RESET_QCOM_STORE_RESET_REASON that is common to both for
exporting the store_reset_reason function and update qcom_wdt_core to
check for this feature instead.

Change-Id: I472dca6757a9463670f3a8b0eea4aa9f6d7fe417
Signed-off-by: Johnathan Cobden-Nolan <quic_jcobdenn@quicinc.com>