fastrpc: Fix to avoid untrusted pointer dereference
Currently, the compat ioctl call distinguishes itself using a global
flag. If a user sends a compat ioctl call followed by a normal ioctl
call, it may result in using a user passed address as a kernel address
in the fastrpcdriver. To address this issue, consider localizing the
compat flag for the ioctl call.

Change-Id: Id72f561d8576d00cb30709ac048de53d5494be73
Signed-off-by: Upendra Chhipa <quic_uchhipa@quicinc.com>