net: stmmac: Add autoneg params from emac partition
Added support to enable/disable C37 auto-negotiation from emac
partition.

New emac partition string could be:
                                             v init/deinit
	"##E6:E0:B7:1D:02:78#255#sgmii#2500##1#1#"
	                                       ^ c37 autoneg
The "1" in the last fields to enable c37 auto-negotion.

Also, this changes device tree property name of C37 auto-negotiation
configuration. The legacy one is to set <1> or <0> to enable/disable C37
AN. Now it change to "qcom,xpcs_autoneg_disabled" to sync with LE1.0.
With ther property set, C37 AN will be disabled.

Change-Id: Ia54aaac5bd056172d40aafb70cfb58c797c2fb5e
Signed-off-by: Henry Yeh <quic_hanyyeh@quicinc.com>