drivers: input: sensors: Fix the irq & CPU consumption issue
When the bottom-half scheduling is delayed, the FIFO length exceeds the
FIFO watermark, causing the IRQ to halt. This results in the absence of
sensor data, an issue predominantly observed after deactivating and
reactivating the sensor.

In some cases, when the sensor is deactivated, the FIFO length does not
reset to zero and remains in the while loop indefinitely. These changes
will exit the while loop when the sensor is moved to the suspend state.

Change-Id: I1d3be3bbb999aa71c38000cc3b77ebbe3d266619
Signed-off-by: Twinkle Goyal <twinkleg@qti.qualcomm.com>